Logic.ly file of the circuit plus 4 screenshots.
On the logic.ly file you have the actual circuit while on the 4 screenshots we have one possibility on each screen shot of each scenario of the failing lamps, white lamp when any 3 are failing, yellow when any 4, orange when any 5 and red when all 6 of them. (All lamps are blue because this is the only color available in logic.ly but each lamp has a label of the color it represents.